Have the Toggle Keyboard Shortcuts Troubleshooting action open the renderer log
